+++ src/main/kotlin/com/christopherosthues/starwarsprogressbar/constants/ConfigurationConstants.kt
@@ -1,5 +1,6 @@
 package com.christopherosthues.starwarsprogressbar.constants
 
+import com.christopherosthues.starwarsprogressbar.configuration.Language
 import com.christopherosthues.starwarsprogressbar.selectors.SelectionType
 
 internal const val DEFAULT_SHOW_VEHICLE = true
@@ -13,3 +14,4 @@ internal const val DEFAULT_DRAW_SILHOUETTES = false
 internal const val DEFAULT_CHANGE_VEHICLE_AFTER_PASS = false
 internal const val DEFAULT_NUMBER_OF_PASSES_UNTIL_VEHICLE_CHANGE = 2
 internal val DEFAULT_VEHICLE_SELECTOR = SelectionType.RANDOM_ALL
+internal val DEFAULT_LANGUAGE = Language.ENGLISH